requirements.txt
Flask==3.1.0
Pillow==11.1.0
pillow-heif==0.21.0
Flask-Cors==5.0.0
